feature: Dry liqudity mining campaign
